obsidian-knowledge-factory/engine/requirements.txt
requests>=2.31.0
PyPDF2>=3.0.0
PyYAML>=6.0
